Change years to int

+++ src/Core/IncomeCalculator/CalculateStrategy/FinalAmountWithRegularPayments.php
@@ -33,7 +33,7 @@ class FinalAmountWithRegularPayments implements CalculateStrategyInterface
             (float)$input->getInitialAmount(),
             (float)$input->getRegularPayment(),
             $input->getNumberOfRegularPaymentsPerYear(),
-            (float)$input->getNumberOfYears(),
+            (int)$input->getNumberOfYears(),
             (float)$input->getInterestRatePerYear(),
             $finalAmount
         );
